+++ bundlers/babel/BabelBundler.js
@@ -1,5 +1,7 @@
 "use strict";
+var stream = require("stream");
 var PathUtil = require("../../utils/PathUtil");
+var BrowserifyHelper = require("../../bundlers/BrowserifyHelper");
 /** Build a JS bundle using the Babel compiler
  * requires package.json:
  *   "babel-core": "~6.18.2",
@@ -7,23 +9,26 @@ var PathUtil = require("../../utils/PathUtil");
  *   "babel-preset-es2015": "~6.18.0",
  *   "babelify": "~7.3.0",
  */
-var BabelBabelify;
-(function (BabelBabelify) {
+var BabelBundler;
+(function (BabelBundler) {
     /** Create a 'browserify' transform which compiles source files using babelify
      */
-    function createTransformer(babelify) {
+    function createTransformer(babelify, filePattern, babelCompilerOpts, transformOpts) {
         var res = {
             transform: function babelifyTransform(tr, opts) {
+                if (filePattern != null && !filePattern.test(tr)) {
+                    return new stream.PassThrough();
+                }
                 console.log("babelify: '" + PathUtil.toShortFileName(tr) + "'");
-                var res = babelify(tr, opts);
-                return res;
+                var strm = babelify(tr, BrowserifyHelper.combineOpts(opts, babelCompilerOpts));
+                return strm;
             },
-            options: {
+            options: BrowserifyHelper.combineOpts({
                 presets: ["es2015"],
-            }
+            }, transformOpts)
         };
         return res;
     }
-    BabelBabelify.createTransformer = createTransformer;
-})(BabelBabelify || (BabelBabelify = {}));
-module.exports = BabelBabelify;
+    BabelBundler.createTransformer = createTransformer;
+})(BabelBundler || (BabelBundler = {}));
+module.exports = BabelBundler;