+++ bundlers/traceur/Es6ifyToStream.js
@@ -4,9 +4,9 @@ var through = require("through");
 /** Modified version of the 'es6ify' library.
  * Works with latest version of traceur and TypeScript
  */
-var Es6ifyLike;
-(function (Es6ifyLike) {
-    Es6ifyLike.traceurOverrides = {};
+var Es6ifyToStream;
+(function (Es6ifyToStream) {
+    Es6ifyToStream.traceurOverrides = {};
     var cache = {};
     var traceurOptions = {
         modules: 'commonjs',
@@ -18,19 +18,13 @@ var Es6ifyLike;
      * file's name, the second argument is true if the file is going to be compiled or false if the file is being skipped/rejected
      * @param dataDone optional callback which is called when a file finishes being compiled
      */
-    function createCompiler(traceur, filePattern, willProcess, dataDone) {
+    function createCompiler(traceur, filePattern, dataDone) {
         filePattern = filePattern || /\.js$/;
         return function es6ifyCompile(file) {
             if (!filePattern.test(file)) {
-                if (willProcess) {
-                    willProcess(file, false);
-                }
                 return through();
             }
             var data = '';
-            if (willProcess) {
-                willProcess(file, true);
-            }
             return through(write, end);
             function write(buf) {
                 data += buf;
@@ -41,7 +35,7 @@ var Es6ifyLike;
                 if (!cached || cached.hash !== hash) {
                     try {
                         cache[file] = {
-                            compiled: compileFile(traceur, file, data, Es6ifyLike.traceurOverrides),
+                            compiled: compileFile(traceur, file, data, Es6ifyToStream.traceurOverrides),
                             hash: hash
                         };
                     }
@@ -58,7 +52,7 @@ var Es6ifyLike;
             }
         };
     }
-    Es6ifyLike.createCompiler = createCompiler;
+    Es6ifyToStream.createCompiler = createCompiler;
     /** Compile function, exposed to be used from other libraries, not needed when using es6ify as a transform.
      * @param {string} file name of the file that is being compiled to ES5
      * @param {string} src source of the file being compiled to ES5
@@ -76,16 +70,16 @@ var Es6ifyLike;
         }
         return result;
     }
-    Es6ifyLike.compileFile = compileFile;
+    Es6ifyToStream.compileFile = compileFile;
     function buildTraceurOptions(overrides) {
         var options = Object.assign({}, traceurOptions, overrides);
         if (typeof options.sourceMap !== 'undefined') {
-            console.warn('es6ify: DEPRECATED traceurOverrides.sourceMap has changed to traceurOverrides.sourceMaps (plural)');
+            console.warn('Es6ifyToStream: DEPRECATED traceurOverrides.sourceMap has changed to traceurOverrides.sourceMaps (plural)');
             options.sourceMaps = options.sourceMap;
             delete options.sourceMap;
         }
         if (options.sourceMaps === true) {
-            console.warn('es6ify: DEPRECATED "traceurOverrides.sourceMaps = true" is not a valid option, traceur sourceMaps options are [false|inline|file]');
+            console.warn('Es6ifyToStream: DEPRECATED "traceurOverrides.sourceMaps = true" is not a valid option, traceur sourceMaps options are [false|inline|file]');
             options.sourceMaps = 'inline';
         }
         return options;
@@ -96,5 +90,5 @@ var Es6ifyLike;
             .update(data)
             .digest('hex');
     }
-})(Es6ifyLike || (Es6ifyLike = {}));
-module.exports = Es6ifyLike;
+})(Es6ifyToStream || (Es6ifyToStream = {}));
+module.exports = Es6ifyToStream;