+++ CHANGELOG.md
@@ -4,7 +4,19 @@ This project does its best to adhere to [Semantic Versioning](http://semver.org/
 
 
 --------
-### [0.2.0](N/A) - 2016-11-08
+### [0.3.0](N/A) - 2016-11-09
+#### Added
+* UglifyBundler and UglifyToStream leveraging 'uglify-js' package to transform files for browserify to bundle
+* Added 'convert-source-map@~1.3.0', 'minimatch@~3.0.3', and 'watchify@~3.7.0' project.json dependencies
+* Added BrowserifyHelper.combineOpts()
+* Added parameters to BabelBundler and TraceurBundler createTransformer()
+
+#### Changed
+* Renamed Es6ifyLike -> Es6ifyToStream
+
+
+--------
+### [0.2.0](https://github.com/TeamworkGuy2/ts-bundlify/commit/6a659706da5ba85a628bada4486870962114b896) - 2016-11-08
 #### Changed
 Complete refactor.
 * Renamed directory util/ -> utils/